minor refactor of api layer models live near endpoints using them
